Questões C, E e F

+++ lista01/C.cpp
@@ -3,9 +3,9 @@
 using namespace std;
 
 int main(){
-    int l1, l2;
+    long long l1, l2;
     cin >> l1 >> l2;
-    int p = (l1*2) + (l2*2);
+    long long p = (l1*2) + (l2*2);
     cout << p;
     return 0; 
 }
